Moved AbstractHorse mounts into core
