+++ v1_14_R1/src/main/java/be/isach/ultracosmetics/v1_14_R1/Module.java
@@ -1,6 +1,9 @@
 package be.isach.ultracosmetics.v1_14_R1;
 
+import be.isach.ultracosmetics.cosmetics.mounts.Mount;
 import be.isach.ultracosmetics.v1_14_R1.customentities.CustomEntities;
+import be.isach.ultracosmetics.v1_14_R1.mount.MountSlime;
+import be.isach.ultracosmetics.v1_14_R1.mount.MountSpider;
 import be.isach.ultracosmetics.version.IModule;
 
 /**
@@ -15,5 +18,15 @@ public class Module implements IModule {
     @Override
     public void disable() {
         CustomEntities.unregisterEntities();
+    }
+
+    @Override
+    public Class<? extends Mount> getSpiderClass() {
+        return MountSpider.class;
+    }
+
+    @Override
+    public Class<? extends Mount> getSlimeClass() {
+        return MountSlime.class;
     }
 }
