+++ v1_8_R3/src/main/java/be/isach/ultracosmetics/v1_8_R3/Module.java
@@ -1,6 +1,9 @@
 package be.isach.ultracosmetics.v1_8_R3;
 
+import be.isach.ultracosmetics.cosmetics.mounts.Mount;
 import be.isach.ultracosmetics.v1_8_R3.customentities.CustomEntities;
+import be.isach.ultracosmetics.v1_8_R3.mount.MountSlime;
+import be.isach.ultracosmetics.v1_8_R3.mount.MountSpider;
 import be.isach.ultracosmetics.version.IModule;
 
 public class Module implements IModule {
@@ -12,5 +15,15 @@ public class Module implements IModule {
     @Override
     public void disable() {
         CustomEntities.unregisterEntities();
+    }
+
+    @Override
+    public Class<? extends Mount> getSpiderClass() {
+        return MountSpider.class;
+    }
+
+    @Override
+    public Class<? extends Mount> getSlimeClass() {
+        return MountSlime.class;
     }
 }
